To Do List
Unable to add null characters to the list. However, it gives the same error when nothing is written.
In the JavaScript code you will write, there must be a function that allows adding elements, a function that allows deleting elements, and a function that allows marking done.
You may have noticed that when the element is added and it gives an error, it gives a warning at the top right. It's the Bootstrap Toast manifest that enables it. Go to the page and research how toast is done and add it to your code.
Go for the main functions first. Done, toast notification these are for some more artistry. It must fulfill its primary purpose.